+++ src/game/Server/Packets/Npc.cpp
@@ -98,13 +98,13 @@ void WorldPackets::Npc::ShowBank::AppendBodyTo(ByteBuffer& buffer) const
 
 void WorldPackets::Npc::StableResult::AppendBodyTo(ByteBuffer& buffer) const
 {
-    buffer << uint8(result);
+    buffer << result;
 }
 
 void WorldPackets::Npc::TrainerBuySucceeded::AppendBodyTo(ByteBuffer& buffer) const
 {
     buffer << ObjectGuid(trainerGuid);
-    buffer << uint32(spellId);
+    buffer << spellId;
 }
 
 void WorldPackets::Npc::TrainerBuyFailed::AppendBodyTo(ByteBuffer& buffer) const
@@ -140,20 +140,20 @@ void WorldPackets::Npc::NpcTextUpdate::AppendBodyTo(ByteBuffer& buffer) const
 void WorldPackets::Npc::TrainerListResponse::AppendBodyTo(ByteBuffer& buffer) const
 {
     buffer << trainerGuid;
-    buffer << uint32(trainerType);
+    buffer << trainerType;
     buffer << uint32(spells.size());
     for (auto const& spell : spells)
     {
-        buffer << uint32(spell.spellId);
-        buffer << uint8(spell.state);
-        buffer << uint32(spell.cost);
-        buffer << uint32(spell.canLearnPrimaryProf);
-        buffer << uint32(spell.isPrimaryProfFirstRank);
-        buffer << uint8(spell.spellLevel);
-        buffer << uint32(spell.reqSkill);
-        buffer << uint32(spell.reqSkillValue);
-        buffer << uint32(spell.spellReq1);
-        buffer << uint32(spell.spellReq2);
+        buffer << spell.spellId;
+        buffer << spell.state;
+        buffer << spell.cost;
+        buffer << spell.canLearnPrimaryProf;
+        buffer << spell.isPrimaryProfFirstRank;
+        buffer << spell.spellLevel;
+        buffer << spell.reqSkill;
+        buffer << spell.reqSkillValue;
+        buffer << spell.spellReq1;
+        buffer << spell.spellReq2;
         buffer << uint32(0);
     }
     buffer << title;
